+++ core/pi_client.py
@@ -5,7 +5,9 @@ from ncatbot.utils		import get_log
 from collections.abc	import AsyncIterator
 from typing				import Optional, Literal
 
-import aiofiles
+from ..errors	import PIPromptBusyError
+
+import asyncio
 
 
 PI_LOGGER = get_log("PiClient")
@@ -13,11 +15,18 @@ PI_LOGGER = get_log("PiClient")
 
 class PiClient(_PiClient):
 	
+	def __init__(self, *args, **kwargs) -> None:
+		
+		super().__init__(*args, **kwargs)
+		
+		# 并发输出安全
+		self._stream_lock: asyncio.Lock	= asyncio.Lock()
+	
 	async def prompt(
 		self,
-		*args,
-		streamingBehavior: Optional[Literal["steer", "followUp"]] = None,
-		**kwargs,
+		message				: str, *,
+		images				: Optional[list]	= None,
+		streamingBehavior	: Optional[Literal["steer", "followUp"]]	= None,
 	) -> AsyncIterator[BaseEvent]:
 		
 		"""
@@ -27,32 +36,31 @@ class PiClient(_PiClient):
 		防止一个事件流被多个 Task 注册导致混乱
 		"""
 		
-		state = await self.get_state()
+		if self._stream_lock.locked():
 		
-		if state.isStreaming is True:
-			
-			PI_LOGGER.debug("pi 正流式输出中...")
+			PI_LOGGER.debug(f"pi 正流式输出中... (streamingBehavior = {streamingBehavior})")
 			
 			if streamingBehavior == "steer":
-				PI_LOGGER.debug("进入 steer 路径")
-				await self.steer(*args, **kwargs)
+				await self.steer(message=message, images=images)
 			
 			elif streamingBehavior == "followUp":
-				PI_LOGGER.debug("进入 follow_up 路径")
-				await self.follow_up(*args, **kwargs)
+				await self.follow_up(message=message, images=images)
 			
-			return # HACK: 防止事件流被多次订阅 但直接return会丢失原因导致另一个订阅者不知道为什么没有收到任何event 后续可能会根据实际情况改成raise ...
+			else:
+				raise PIPromptBusyError()
+			
+			return
+		
+		await self._stream_lock.acquire()
 		
 		try:
-			async for event in super().prompt(*args, **kwargs):
+			
+			async for event in super().prompt(message=message, images=images):
 				yield event
 		
-		except ValueError as e:
-			
-			PI_LOGGER.error(f"prompt 出错: {e}")
-			
-			buffer: bytes = await self._transport._io._process.stdout.read(1024 * 1024 * 32)
-			async with aiofiles.open("/root/lastest_crash.log", "wb") as file:
-				await file.write(buffer)
-			
+		except Exception as e:
+			PI_LOGGER.exception(f"prompt 出错: {e}")
 			raise
+		
+		finally:
+			self._stream_lock.release()