+++ errors/base_bot_error.py
@@ -1,0 +1,3 @@
+class BaseBotError(Exception):
+	"""机器人基类错误"""
+	...